+++ src/main/kotlin/com/github/bytecodealliance/language/mixin/MixinFlags.kt
@@ -18,7 +18,7 @@ abstract class MixinFlags(node: ASTNode) : WitElement(node),
     }
 
     override fun getNameIdentifier(): PsiElement? {
-        return this.identifierSafe
+        return this.identifierFree
     }