+++ CHANGELOG.md
@@ -5,8 +5,11 @@
 ## [0.1.2]
 ### Added
 - Support `@since` annotation
+- Support mark `getter` function
+- Support mark `setter` function
 ### Fix
 - Remove `Nested namespace dost not support yet` hint
+- Fix `world` indent level
 
 ## [0.1.1]
 ### Fix